[adone] refactor to module style
