+++ types/adone/test/glosses/std.ts
@@ -1,3 +1,5 @@
+import adone from "adone";
+
 import * as assert from "assert";
 import * as fs from "fs";
 import * as path from "path";
